datastore/requirements.txt
Flask>=1.0.0
gunicorn==19.7.1
google-cloud-datastore==1.4.0
google-cloud-storage==1.7.0
oauth2client==4.1.2
six==1.11.0